Refactor PDF to Excel converter by moving the class implementation to a separate module; update main script to utilize the new module and remove redundant code; delete obsolete test notebook.
